separate header for zoom image and make sure title no overlap close button
