+++ libs/expo/shared/ui-components/src/lib/ImagesWithZoom/ImagesWithZoom.tsx
@@ -1,10 +1,8 @@
-import { PlusIcon } from '@monorepo/expo/shared/icons';
-import { Colors, Spacings } from '@monorepo/expo/shared/static';
+import { Colors } from '@monorepo/expo/shared/static';
 import { useState } from 'react';
-import { Modal, Pressable, View } from 'react-native';
+import { Modal, Pressable } from 'react-native';
 import ImageViewer from 'react-native-image-zoom-viewer';
-import { useSafeAreaInsets } from 'react-native-safe-area-context';
-import TextMedium from '../TextMedium';
+import ImagesWithZoomHeader from './ImagesWithZoomHeader';
 
 interface ImagesWithZoomProps {
   title?: string | null;
@@ -15,8 +13,6 @@ interface ImagesWithZoomProps {
 export function ImagesWithZoom(props: ImagesWithZoomProps) {
   const { title, imageUrls, children } = props;
   const [isVisible, setIsVisible] = useState(false);
-
-  const insets = useSafeAreaInsets();
 
   return (
     <>
@@ -33,35 +29,7 @@ export function ImagesWithZoom(props: ImagesWithZoomProps) {
         }}
         visible={isVisible}
       >
-        <Pressable
-          onPress={() => setIsVisible(false)}
-          style={{
-            position: 'absolute',
-            top: insets.top + Spacings.sm,
-            right: Spacings.xxs,
-            height: 40,
-            width: 40,
-            zIndex: 2,
-          }}
-          accessibilityHint="close"
-          accessibilityRole="button"
-        >
-          <PlusIcon size="md" color={Colors.BLACK} rotate="45deg" />
-        </Pressable>
-        {title && (
-          <View
-            style={{
-              position: 'absolute',
-              top: insets.top + Spacings.sm,
-              left: 0,
-              right: 0,
-              alignItems: 'center',
-              zIndex: 1,
-            }}
-          >
-            <TextMedium size="lg">{title}</TextMedium>
-          </View>
-        )}
+        <ImagesWithZoomHeader title={title} setIsVisible={setIsVisible} />
         <ImageViewer
           backgroundColor="white"
           renderIndicator={() => <></>}
